+++ tsdb/compact.go
@@ -29,7 +29,6 @@ import (
 	"github.com/oklog/ulid"
 	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/prometheus/pkg/labels"
 	"github.com/prometheus/prometheus/storage"
 	"github.com/prometheus/prometheus/tsdb/chunkenc"
 	"github.com/prometheus/prometheus/tsdb/chunks"
@@ -786,398 +785,4 @@ func (c *LeveledCompactor) populateBlock(blocks []BlockReader, meta *BlockMeta, 
 	}
 
 	return nil
-}
-
-// blockBaseSeriesSet allows to iterate over all series in the single block.
-// Iterated series are trimmed with given min and max time as well as tombstones.
-// See newBlockSeriesSet and newBlockChunkSeriesSet to use it for either sample or chunk iterating.
-type blockBaseSeriesSet struct {
-	p          index.Postings
-	index      IndexReader
-	chunks     ChunkReader
-	tombstones tombstones.Reader
-	mint, maxt int64
-
-	currIterFn func() *populateWithDelGenericSeriesIterator
-	currLabels labels.Labels
-
-	bufChks []chunks.Meta
-	err     error
-}
-
-func (b *blockBaseSeriesSet) Next() bool {
-	var lbls labels.Labels
-
-	for b.p.Next() {
-		if err := b.index.Series(b.p.At(), &lbls, &b.bufChks); err != nil {
-			// Postings may be stale. Skip if no underlying series exists.
-			if errors.Cause(err) == storage.ErrNotFound {
-				continue
-			}
-			b.err = errors.Wrapf(err, "get series %d", b.p.At())
-			return false
-		}
-
-		if len(b.bufChks) == 0 {
-			continue
-		}
-
-		intervals, err := b.tombstones.Get(b.p.At())
-		if err != nil {
-			b.err = errors.Wrap(err, "get tombstones")
-			return false
-		}
-
-		// NOTE:
-		// * block time range is half-open: [meta.MinTime, meta.MaxTime).
-		// * chunks are both closed: [chk.MinTime, chk.MaxTime].
-		// * requested time ranges are closed: [req.Start, req.End].
-
-		var trimFront, trimBack bool
-
-		// Copy chunks as iteratables are reusable.
-		chks := make([]chunks.Meta, 0, len(b.bufChks))
-
-		// Prefilter chunks and pick those which are not entirely deleted or totally outside of the requested range.
-		for _, chk := range b.bufChks {
-			if chk.MaxTime < b.mint {
-				continue
-			}
-			if chk.MinTime > b.maxt {
-				continue
-			}
-
-			if !(tombstones.Interval{Mint: chk.MinTime, Maxt: chk.MaxTime}.IsSubrange(intervals)) {
-				chks = append(chks, chk)
-			}
-
-			// If still not entirely deleted, check if trim is needed based on requested time range.
-			if chk.MinTime < b.mint {
-				trimFront = true
-			}
-			if chk.MaxTime > b.maxt {
-				trimBack = true
-			}
-		}
-
-		if len(chks) == 0 {
-			continue
-		}
-
-		if trimFront {
-			intervals = intervals.Add(tombstones.Interval{Mint: math.MinInt64, Maxt: b.mint - 1})
-		}
-		if trimBack {
-			intervals = intervals.Add(tombstones.Interval{Mint: b.maxt + 1, Maxt: math.MaxInt64})
-		}
-		b.currLabels = lbls
-		b.currIterFn = func() *populateWithDelGenericSeriesIterator {
-			return newPopulateWithDelGenericSeriesIterator(b.chunks, chks, intervals)
-		}
-		return true
-	}
-	return false
-}
-
-func (b *blockBaseSeriesSet) Err() error {
-	if b.err != nil {
-		return b.err
-	}
-	return b.p.Err()
-}
-
-func (b *blockBaseSeriesSet) Warnings() storage.Warnings { return nil }
-
-// populateWithDelGenericSeriesIterator allows to iterate over given chunk metas. In each iteration it ensures
-// that chunks are trimmed based on given tombstones interval if any.
-//
-// populateWithDelGenericSeriesIterator assumes that chunks that would be fully removed by intervals are filtered out in previous phase.
-//
-// On each iteration currChkMeta is available. If currDelIter is not nil, it means that chunk iterator in currChkMeta
-// is invalid and chunk rewrite is needed, currDelIter should be used.
-type populateWithDelGenericSeriesIterator struct {
-	chunks ChunkReader
-	// chks are expected to be sorted by minTime and should be related to the same, single series.
-	chks []chunks.Meta
-
-	i         int
-	err       error
-	bufIter   *deletedIterator
-	intervals tombstones.Intervals
-
-	currDelIter chunkenc.Iterator
-	currChkMeta chunks.Meta
-}
-
-func newPopulateWithDelGenericSeriesIterator(
-	chunks ChunkReader,
-	chks []chunks.Meta,
-	intervals tombstones.Intervals,
-) *populateWithDelGenericSeriesIterator {
-	return &populateWithDelGenericSeriesIterator{
-		chunks:    chunks,
-		chks:      chks,
-		i:         -1,
-		bufIter:   &deletedIterator{},
-		intervals: intervals,
-	}
-}
-
-func (p *populateWithDelGenericSeriesIterator) next() bool {
-	if p.err != nil || p.i >= len(p.chks)-1 {
-		return false
-	}
-
-	p.i++
-	p.currChkMeta = p.chks[p.i]
-
-	p.currChkMeta.Chunk, p.err = p.chunks.Chunk(p.currChkMeta.Ref)
-	if p.err != nil {
-		p.err = errors.Wrapf(p.err, "cannot populate chunk %d", p.currChkMeta.Ref)
-		return false
-	}
-
-	p.bufIter.intervals = p.bufIter.intervals[:0]
-	for _, interval := range p.intervals {
-		if p.currChkMeta.OverlapsClosedInterval(interval.Mint, interval.Maxt) {
-			p.bufIter.intervals = p.bufIter.intervals.Add(interval)
-		}
-	}
-
-	// Re-encode head chunks that are still open (being appended to) or
-	// outside the compacted MaxTime range.
-	// The chunk.Bytes() method is not safe for open chunks hence the re-encoding.
-	// This happens when snapshotting the head block or just fetching chunks from TSDB.
-	//
-	// TODO think how to avoid the typecasting to verify when it is head block.
-	_, isSafeChunk := p.currChkMeta.Chunk.(*safeChunk)
-	if len(p.bufIter.intervals) == 0 && !(isSafeChunk && p.currChkMeta.MaxTime == math.MaxInt64) {
-		// If there are no overlap with deletion intervals AND it's NOT an "open" head chunk, we can take chunk as it is.
-		p.currDelIter = nil
-		return true
-	}
-
-	// We don't want full chunk or it's potentially still opened, take just part of it.
-	p.bufIter.it = p.currChkMeta.Chunk.Iterator(nil)
-	p.currDelIter = p.bufIter
-	return true
-}
-
-func (p *populateWithDelGenericSeriesIterator) Err() error { return p.err }
-
-func (p *populateWithDelGenericSeriesIterator) toSeriesIterator() chunkenc.Iterator {
-	return &populateWithDelSeriesIterator{populateWithDelGenericSeriesIterator: p}
-}
-func (p *populateWithDelGenericSeriesIterator) toChunkSeriesIterator() chunks.Iterator {
-	return &populateWithDelChunkSeriesIterator{populateWithDelGenericSeriesIterator: p}
-}
-
-// populateWithDelSeriesIterator allows to iterate over samples for the single series.
-type populateWithDelSeriesIterator struct {
-	*populateWithDelGenericSeriesIterator
-
-	curr chunkenc.Iterator
-}
-
-func (p *populateWithDelSeriesIterator) Next() bool {
-	if p.curr != nil && p.curr.Next() {
-		return true
-	}
-
-	for p.next() {
-		if p.currDelIter != nil {
-			p.curr = p.currDelIter
-		} else {
-			p.curr = p.currChkMeta.Chunk.Iterator(nil)
-		}
-		if p.curr.Next() {
-			return true
-		}
-	}
-	return false
-}
-
-func (p *populateWithDelSeriesIterator) Seek(t int64) bool {
-	if p.curr != nil && p.curr.Seek(t) {
-		return true
-	}
-	for p.Next() {
-		if p.curr.Seek(t) {
-			return true
-		}
-	}
-	return false
-}
-
-func (p *populateWithDelSeriesIterator) At() (int64, float64) { return p.curr.At() }
-
-func (p *populateWithDelSeriesIterator) Err() error {
-	if err := p.populateWithDelGenericSeriesIterator.Err(); err != nil {
-		return err
-	}
-	if p.curr != nil {
-		return p.curr.Err()
-	}
-	return nil
-}
-
-type populateWithDelChunkSeriesIterator struct {
-	*populateWithDelGenericSeriesIterator
-
-	curr chunks.Meta
-}
-
-func (p *populateWithDelChunkSeriesIterator) Next() bool {
-	if !p.next() {
-		return false
-	}
-
-	p.curr = p.currChkMeta
-	if p.currDelIter == nil {
-		return true
-	}
-
-	// Re-encode the chunk if iterator is provider. This means that it has some samples to be deleted or chunk is opened.
-	newChunk := chunkenc.NewXORChunk()
-	app, err := newChunk.Appender()
-	if err != nil {
-		p.err = err
-		return false
-	}
-
-	if !p.currDelIter.Next() {
-		if err := p.currDelIter.Err(); err != nil {
-			p.err = errors.Wrap(err, "iterate chunk while re-encoding")
-			return false
-		}
-
-		// Empty chunk, this should not happen, as we assume full deletions being filtered before this iterator.
-		p.err = errors.Wrap(err, "populateWithDelChunkSeriesIterator: unexpected empty chunk found while rewriting chunk")
-		return false
-	}
-
-	t, v := p.currDelIter.At()
-	p.curr.MinTime = t
-	app.Append(t, v)
-
-	for p.currDelIter.Next() {
-		t, v = p.currDelIter.At()
-		app.Append(t, v)
-	}
-	if err := p.currDelIter.Err(); err != nil {
-		p.err = errors.Wrap(err, "iterate chunk while re-encoding")
-		return false
-	}
-
-	p.curr.Chunk = newChunk
-	p.curr.MaxTime = t
-	return true
-}
-
-func (p *populateWithDelChunkSeriesIterator) At() chunks.Meta { return p.curr }
-
-// blockSeriesSet allows to iterate over sorted, populated series with applied tombstones.
-// Series with all deleted chunks are still present as Series with no samples.
-// Samples from chunks are also trimmed to requested min and max time.
-type blockSeriesSet struct {
-	blockBaseSeriesSet
-}
-
-func newBlockSeriesSet(i IndexReader, c ChunkReader, t tombstones.Reader, p index.Postings, mint, maxt int64) storage.SeriesSet {
-	return &blockSeriesSet{
-		blockBaseSeriesSet{
-			index:      i,
-			chunks:     c,
-			tombstones: t,
-			p:          p,
-			mint:       mint,
-			maxt:       maxt,
-		},
-	}
-}
-
-func (b *blockSeriesSet) At() storage.Series {
-	// At can be looped over before iterating, so save the current value locally.
-	currIterFn := b.currIterFn
-	return &storage.SeriesEntry{
-		Lset: b.currLabels,
-		SampleIteratorFn: func() chunkenc.Iterator {
-			return currIterFn().toSeriesIterator()
-		},
-	}
-}
-
-// blockChunkSeriesSet allows to iterate over sorted, populated series with applied tombstones.
-// Series with all deleted chunks are still present as Labelled iterator with no chunks.
-// Chunks are also trimmed to requested [min and max] (keeping samples with min and max timestamps).
-type blockChunkSeriesSet struct {
-	blockBaseSeriesSet
-}
-
-func newBlockChunkSeriesSet(i IndexReader, c ChunkReader, t tombstones.Reader, p index.Postings, mint, maxt int64) storage.ChunkSeriesSet {
-	return &blockChunkSeriesSet{
-		blockBaseSeriesSet{
-			index:      i,
-			chunks:     c,
-			tombstones: t,
-			p:          p,
-			mint:       mint,
-			maxt:       maxt,
-		},
-	}
-}
-
-func (b *blockChunkSeriesSet) At() storage.ChunkSeries {
-	// At can be looped over before iterating, so save the current value locally.
-	currIterFn := b.currIterFn
-	return &storage.ChunkSeriesEntry{
-		Lset: b.currLabels,
-		ChunkIteratorFn: func() chunks.Iterator {
-			return currIterFn().toChunkSeriesIterator()
-		},
-	}
-}
-
-func newMergedStringIter(a index.StringIter, b index.StringIter) index.StringIter {
-	return &mergedStringIter{a: a, b: b, aok: a.Next(), bok: b.Next()}
-}
-
-type mergedStringIter struct {
-	a        index.StringIter
-	b        index.StringIter
-	aok, bok bool
-	cur      string
-}
-
-func (m *mergedStringIter) Next() bool {
-	if (!m.aok && !m.bok) || (m.Err() != nil) {
-		return false
-	}
-
-	if !m.aok {
-		m.cur = m.b.At()
-		m.bok = m.b.Next()
-	} else if !m.bok {
-		m.cur = m.a.At()
-		m.aok = m.a.Next()
-	} else if m.b.At() > m.a.At() {
-		m.cur = m.a.At()
-		m.aok = m.a.Next()
-	} else if m.a.At() > m.b.At() {
-		m.cur = m.b.At()
-		m.bok = m.b.Next()
-	} else { // Equal.
-		m.cur = m.b.At()
-		m.aok = m.a.Next()
-		m.bok = m.b.Next()
-	}
-
-	return true
-}
-func (m mergedStringIter) At() string { return m.cur }
-func (m mergedStringIter) Err() error {
-	if m.a.Err() != nil {
-		return m.a.Err()
-	}
-	return m.b.Err()
 }
